refactor: config 디렉터리 분
