+++ src/main/java/org/phoenix/planet/configuration/security/CorsConfig.java
@@ -1,13 +1,12 @@
-package org.phoenix.planet.configuration;
+package org.phoenix.planet.configuration.security;
 
 
+import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import java.util.List;
 
 @Configuration
 public class CorsConfig {
@@ -18,25 +17,25 @@ public class CorsConfig {
         CorsConfiguration config = new CorsConfiguration();
 
         config.setAllowedOriginPatterns(List.of(
-                "http://localhost:5173",
-                "http://127.0.0.1:5173",
-                "https://planet-frontend.jaeyoung.store"
+            "http://localhost:5173",
+            "http://127.0.0.1:5173",
+            "https://planet-frontend.jaeyoung.store"
         ));
 //        config.setAllowedOriginPatterns(List.of("*"));
         config.setAllowedMethods(List.of(
-                "GET",
-                "POST",
-                "PUT",
-                "PATCH",
-                "DELETE",
-                "OPTIONS"
+            "GET",
+            "POST",
+            "PUT",
+            "PATCH",
+            "DELETE",
+            "OPTIONS"
         ));
         config.setAllowedHeaders(List.of("*"));
         config.setExposedHeaders(List.of(
-                "Authorization",
-                "Set-Cookie",
-                "Location",
-                "X-Error-Code"
+            "Authorization",
+            "Set-Cookie",
+            "Location",
+            "X-Error-Code"
         ));
         config.setAllowCredentials(true);
 //        config.setMaxAge(Duration.ofHours(1)); // 뭐임?
