+++ src/main/java/org/phoenix/planet/configuration/security/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package org.phoenix.planet.configuration;
+package org.phoenix.planet.configuration.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
