+++ src/main/java/org/phoenix/planet/configuration/security/OAuth2SuccessHandler.java
@@ -1,4 +1,4 @@
-package org.phoenix.planet.configuration;
+package org.phoenix.planet.configuration.security;
 
 
 import jakarta.servlet.http.HttpServletRequest;
