+++ src/main/java/org/phoenix/planet/configuration/security/UrlRewriteFilter.java
@@ -1,22 +1,21 @@
-package org.phoenix.planet.configuration;
+package org.phoenix.planet.configuration.security;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletRequestWrapper;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
 
 @Component
 public class UrlRewriteFilter extends OncePerRequestFilter {
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
-                                    FilterChain filterChain)
-            throws ServletException, IOException {
+        FilterChain filterChain)
+        throws ServletException, IOException {
 
         String path = request.getRequestURI();
         if (path.startsWith("/api/")) {
